+++ src/components/Hotels.jsx
@@ -1,9 +1,8 @@
 import React, { useState } from "react";
 import { connect } from "react-redux";
 import { Button } from "semantic-ui-react";
-import { addHotels } from "../modules/destination.js";
+import { addHotels, objectEraser } from "../modules/destination.js";
 import { sliderChoice } from "../helpers/methods.js";
-import { objectEraser } from "../helpers/methods.js";
 
 const Hotels = props => {
   const [hotelBudget, setHotelBudget] = useState(null);
@@ -26,7 +25,7 @@ const Hotels = props => {
   return (
     <>
       {props.message} Let's move on to...
-      <Button
+      <Button id="back-button-4"
         onClick={async () => {
           await objectEraser("activity_types", props.trip, "resturant");
           props.updateProgression(props.progression - 2);
