+++ src/components/Restaurants.jsx
@@ -1,9 +1,8 @@
 import React, { useState } from "react";
 import { connect } from "react-redux";
 import { Dropdown, Button } from "semantic-ui-react";
-import { addRestaurants } from "../modules/destination.js";
+import { addRestaurants, objectEraser } from "../modules/destination.js";
 import { sliderChoice } from "../helpers/methods.js";
-import { objectEraser } from "../helpers/methods.js";
 
 const Trip = props => {
   const [foodBudget, setFoodBudget] = useState(null);
@@ -48,13 +47,13 @@ const Trip = props => {
   return (
     <>
       {props.message} We move on to...
-      <Button
+      <Button id="back-button-5"
         onClick={async () => {
           await objectEraser("hotels", props.trip);
           props.updateProgression(props.progression - 1);
         }}
       >
-        Back a step
+        Back one step
       </Button>
       <div className="food-choice">
         <h2>Food preference:</h2>
