+++ src/components/Trip.jsx
@@ -1,4 +1,4 @@
-import React, { useState} from "react";
+import React from "react";
 import { connect } from "react-redux";
 import Restaurants from "./Restaurants";
 import Hotels from "./Hotels";
