+++ src/helpers/methods.js
@@ -1,5 +1,3 @@
-import axios from "axios";
-
 const sliderChoice = event => {
   Array.from(
     document.getElementsByClassName(`range-values-${event.target.name}`)[0]
@@ -13,17 +11,4 @@ const sliderChoice = event => {
   }
 };
 
-const objectEraser = async (component, trip, restaurant) => {
-  try {
-    const response = await axios({
-      method: "DELETE",
-      url: `api/v1/${component}`,
-      params: { trip: trip, activity_type: restaurant }
-    });
-    return response;
-  } catch (error) {
-    return error;
-  }
-};
-
-export { sliderChoice, objectEraser };
+export { sliderChoice };
